Add basic functionality: add/delete rows via selecting each property
Adding a new row is now done by selecting values in the selectorRow and then submitting them.

The current values of the selectorRow are represented in the model, and when the user selects different values, these updates are registered in the model.
Each field (audioConfig, diameter, thickness, speed) is updated via a different message. At first, I was using just one message covering all of them, and then casing within. This ended up being super messy though. It's still imperfect, but much cleaner to have separate messages and then recuperate the field type from the string passed by the onChange FN.
So now, when a user clicks 'Add new row', the selectorRow values are plucked out of the model and used to create the new row.

New rows can be deleted via a button on the right.

Still to be implemented:
* quantity
* the info (direction, # of passes) that the audioConfig implies
* the actual length calculation for each row
* the total calculation.