Rewrite translation file
Now, some AppStrings take arguments and deliver back a full phrase, rather than the in-View stitching together that was previously happening. The first information paragraph, containing all the links, has now been factored out into a separate FN in the translation file.
The translation file is also formatted to be a bit nicer.